build(phpunit): separates phpunit from my base dependencies
