+++ composer.json
@@ -16,7 +16,6 @@
     "require": {
         "php": "^8.2|^8.3|^8.4",
         "nikic/php-parser": "^5.4",
-        "phpunit/phpunit": "^12.0",
         "symfony/console": "^7.2",
         "symfony/filesystem": "^7.2",
         "symfony/finder": "^7.2"
@@ -26,6 +25,7 @@
         "phpstan/phpstan": "^2.1",
         "phpstan/phpstan-phpunit": "^2.0",
         "phpstan/phpstan-strict-rules": "^2.0",
+        "phpunit/phpunit": "^12.1",
         "rector/rector": "^2.0",
         "shipmonk/composer-dependency-analyser": "^1.8"
     },
