+++ tests/Unit/Asserts/AndTest.php
@@ -6,11 +6,10 @@ namespace Structura\Tests\Unit\Asserts;
 
 use ArrayAccess;
 use Iterator;
-use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Structura\Expr;
-use Structura\Testing\ArchitectureAsserts;
+use Structura\Tests\Helper\ArchitectureAsserts;
 
 class AndTest extends TestCase
 {
