+++ tests/Unit/Asserts/ToBeAnonymousClassesTest.php
@@ -11,7 +11,7 @@ use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Structura\Asserts\ToBeAnonymousClasses;
 use Structura\Expr;
-use Structura\Testing\ArchitectureAsserts;
+use Structura\Tests\Helper\ArchitectureAsserts;
 
 #[CoversClass(ToBeAnonymousClasses::class)]
 class ToBeAnonymousClassesTest extends TestCase
