+++ tests/Unit/Asserts/ToBeClassesTest.php
@@ -9,10 +9,9 @@ use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
-use Structura\Asserts\ToBeAnonymousClasses;
 use Structura\Asserts\ToBeClasses;
 use Structura\Expr;
-use Structura\Testing\ArchitectureAsserts;
+use Structura\Tests\Helper\ArchitectureAsserts;
 
 #[CoversClass(ToBeClasses::class)]
 class ToBeClassesTest extends TestCase
