+++ tests/Unit/Asserts/ToUseNothingTest.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Structura\Expr;
-use Structura\Testing\ArchitectureAsserts;
+use Structura\Tests\Helper\ArchitectureAsserts;
 use Structura\ValueObjects\ExpectValueObject;
 
 class ToUseNothingTest extends TestCase
