+++ tests/Unit/Asserts/ToUseStrictTypesTest.php
@@ -7,7 +7,7 @@ namespace Structura\Tests\Unit\Asserts;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Structura\Expr;
-use Structura\Testing\ArchitectureAsserts;
+use Structura\Tests\Helper\ArchitectureAsserts;
 
 class ToUseStrictTypesTest extends TestCase
 {
